♻️ extract createNewEvent
